# api/docs/swagger.yaml
basePath: /
definitions:
...
  errors.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      path:
        type: string
      request_id:
        type: string
      retry_after:
        type: integer
      timestamp:
        type: string
    type: object
  errors.Links:
    properties:
      first:
        type: string
      last:
        type: string
      next:
        type: string
      prev:
        type: string
      self:
        type: string
    type: object
  errors.Meta:
    properties:
      links:
        $ref: '#/definitions/errors.Links'
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      timestamp:
        type: string
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  errors.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/errors.ErrorInfo'
      meta:
        $ref: '#/definitions/errors.Meta'
      success:
        type: boolean
    type: object
...
  sliders.CreateSliderItemRequest:
    properties:
      content:
        maxLength: 1000
        type: string
      image_url:
        maxLength: 2048
        minLength: 1
        type: string
      link_url:
        maxLength: 2048
        type: string
      order:
        minimum: 0
        type: integer
      tags:
        items:
          type: string
        type: array
      titulo:
        maxLength: 255
        type: string
    required:
    - image_url
    - order
    type: object
  sliders.CreateSliderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/sliders.CreateSliderItemRequest'
        type: array
      location:
        maxLength: 255
        minLength: 1
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
      type:
        maximum: 2
        minimum: 0
        type: integer
    required:
    - location
    - name
    - type
    type: object
  sliders.SliderItemResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      link_url:
        type: string
      order:
        type: integer
      slider_id:
        type: integer
      tags:
        items:
          type: string
        type: array
      titulo:
        type: string
      updated_at:
        type: string
    type: object
  sliders.SliderResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/sliders.SliderItemResponse'
        type: array
      location:
        type: string
      name:
        type: string
      type:
        type: integer
      updated_at:
        type: string
    type: object
  sliders.UpdateSliderItemRequest:
    properties:
      content:
        maxLength: 1000
        type: string
      image_url:
        maxLength: 2048
        minLength: 1
        type: string
      link_url:
        maxLength: 2048
        type: string
      order:
        minimum: 0
        type: integer
      tags:
        items:
          type: string
        type: array
      titulo:
        maxLength: 255
        type: string
    type: object
  sliders.UpdateSliderRequest:
    properties:
      location:
        maxLength: 255
        minLength: 1
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
      type:
        maximum: 2
        minimum: 0
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A production-ready REST API boilerplate in Go with JWT authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: TRIIIO API Boilerplate
  version: "1.0"
paths:
...
  /api/v1/sliders:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of sliders
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/sliders.SliderResponse'
                  type: array
                success:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      summary: List sliders
      tags:
      - sliders
    post:
      consumes:
      - application/json
      description: Create a new slider with items
      parameters:
      - description: Slider creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sliders.CreateSliderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  $ref: '#/definitions/sliders.SliderResponse'
                success:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Create slider
      tags:
      - sliders
  /api/v1/sliders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a slider and all its items
      parameters:
      - description: Slider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Delete slider
      tags:
      - sliders
    get:
      consumes:
      - application/json
      description: Retrieve a slider and its items by ID
      parameters:
      - description: Slider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  $ref: '#/definitions/sliders.SliderResponse'
                success:
                  type: boolean
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      summary: Get slider by ID
      tags:
      - sliders
    put:
      consumes:
      - application/json
      description: Update an existing slider
      parameters:
      - description: Slider ID
        in: path
        name: id
        required: true
        type: integer
      - description: Slider update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sliders.UpdateSliderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  $ref: '#/definitions/sliders.SliderResponse'
                success:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Update slider
      tags:
      - sliders
  /api/v1/sliders/{slider_id}/items:
    get:
      consumes:
      - application/json
      description: Retrieve all items for a specific slider
      parameters:
      - description: Slider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/sliders.SliderItemResponse'
                  type: array
                success:
                  type: boolean
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      summary: Get slider items
      tags:
      - sliders
    post:
      consumes:
      - application/json
      description: Add a new item to an existing slider
      parameters:
      - description: Slider ID
        in: path
        name: id
        required: true
        type: integer
      - description: Slider item creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sliders.CreateSliderItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  $ref: '#/definitions/sliders.SliderItemResponse'
                success:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Add slider item
      tags:
      - sliders
  /api/v1/sliders/items/{item_id}:
    delete:
      consumes:
      - application/json
      description: Delete a slider item
      parameters:
      - description: Slider item ID
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Delete slider item
      tags:
      - sliders
    get:
      consumes:
      - application/json
      description: Retrieve a specific slider item by ID
      parameters:
      - description: Slider item ID
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  $ref: '#/definitions/sliders.SliderItemResponse'
                success:
                  type: boolean
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      summary: Get slider item
      tags:
      - sliders
    put:
      consumes:
      - application/json
      description: Update an existing slider item
      parameters:
      - description: Slider item ID
        in: path
        name: item_id
        required: true
        type: integer
      - description: Slider item update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sliders.UpdateSliderItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  $ref: '#/definitions/sliders.SliderItemResponse'
                success:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Update slider item
      tags:
      - sliders
  /api/v1/sliders/location:
    get:
      consumes:
      - application/json
      description: Retrieve a slider and its items by location
      parameters:
      - description: Slider location
        in: query
        name: location
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                data:
                  $ref: '#/definitions/sliders.SliderResponse'
                success:
                  type: boolean
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/errors.Response'
            - properties:
                error:
                  $ref: '#/definitions/errors.ErrorInfo'
                success:
                  type: boolean
              type: object
      summary: Get slider by location
      tags:
      - sliders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"